+++ s4_viajes-back/src/main/java/co/edu/uniandes/csw/viajes/entities/TransporteTerrestreEntity.java
@@ -28,13 +28,13 @@ public class TransporteTerrestreEntity extends BaseEntity implements Serializabl
 
     private String destino;
 
-    private Long latitudOrigen;
+    private double latitudOrigen;
 
-    private Long longitudOrigen;
+    private double longitudOrigen;
 
-    private Long latitudDestino;
+    private double latitudDestino;
 
-    private Long longitudDestino;
+    private double longitudDestino;
 
     private Integer numeroDias;
 
@@ -79,35 +79,35 @@ public class TransporteTerrestreEntity extends BaseEntity implements Serializabl
         this.destino = destino;
     }
 
-    public Long getLatitudOrigen() {
+    public double getLatitudOrigen() {
         return latitudOrigen;
     }
 
-    public void setLatitudOrigen(Long latitudOrigen) {
+    public void setLatitudOrigen(double latitudOrigen) {
         this.latitudOrigen = latitudOrigen;
     }
 
-    public Long getLongitudOrigen() {
+    public double getLongitudOrigen() {
         return longitudOrigen;
     }
 
-    public void setLongitudOrigen(Long longitudOrigen) {
+    public void setLongitudOrigen(double longitudOrigen) {
         this.longitudOrigen = longitudOrigen;
     }
 
-    public Long getLatitudDestino() {
+    public double getLatitudDestino() {
         return latitudDestino;
     }
 
-    public void setLatitudDestino(Long latitudDestino) {
+    public void setLatitudDestino(double latitudDestino) {
         this.latitudDestino = latitudDestino;
     }
 
-    public Long getLongitudDestino() {
+    public double getLongitudDestino() {
         return longitudDestino;
     }
 
-    public void setLongitudDestino(Long longitudDestino) {
+    public void setLongitudDestino(double longitudDestino) {
         this.longitudDestino = longitudDestino;
     }
